+++ pkg/scheduler/static.go
@@ -1,0 +1,23 @@
+package scheduler
+
+// StaticScheduler definition
+type StaticScheduler struct {
+	job func()
+}
+
+// NewStaticScheduler to create new static scheduler
+func NewStaticScheduler() *StaticScheduler {
+	return &StaticScheduler{}
+}
+
+// AddFunc to add new function to cron scheduler
+func (ss *StaticScheduler) AddFunc(spec string, cmd func()) error {
+	ss.job = cmd
+	return nil
+}
+
+// Start to start the cron scheduler
+func (ss *StaticScheduler) Start() error {
+	ss.job()
+	return nil
+}
